move declarations in header file
